requirements.txt
djangorestframework
django-cors-headers
django-cleanup
pillow
djoser
pytils
django-ckeditor
django
python-pdf
pyfcm

celery
redis

yookassa

flask
python-telegram-bot==13.15